+++ ChangeLog.md
@@ -14,6 +14,7 @@ and this project adheres to
 ### Added
 
 - 24-bit color support to Excel builder
+- cell border support to Excel builder